refactor so that the route rendering component is in its own file
